Add user to Product Rating Entity

+++ src/products/product-ratings/product-ratings.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   ParseIntPipe,
@@ -11,6 +12,10 @@ import {
 import { ProductRatingsService } from './product-ratings.service';
 import { ProductRatingDto } from '../dto/product-rating.dto';
 import { ProductRating } from '../entities/product-rating.entity';
+import { Roles } from '../../auth/roles.decorator';
+import { Role } from '../../users/entities/role.enum';
+import { ReqUser } from '../../auth/user.decorator';
+import { User } from '../../users/entities/user.entity';
 
 @Controller('product-ratings')
 export class ProductRatingsController {
@@ -24,22 +29,34 @@ export class ProductRatingsController {
   }
 
   @Post(':productId')
+  @Roles(Role.Admin, Role.Manager, Role.Sales, Role.Customer)
   async createProductRating(
+    @ReqUser() user: User,
     @Param('productId', ParseIntPipe) productId: number,
     @Body() body: ProductRatingDto,
   ): Promise<ProductRating> {
     return await this.productRatingsService.createProductRating(
+      user,
       productId,
       body,
     );
   }
 
   @Put(':productId/:id')
+  @Roles(Role.Admin, Role.Manager, Role.Sales, Role.Customer)
   async updateProductRating(
+    @ReqUser() user: User,
     @Param('productId', ParseIntPipe) productId: number,
     @Param('id', ParseIntPipe) id: number,
     @Body() body: ProductRatingDto,
   ): Promise<ProductRating> {
+    const checkUser = await this.productRatingsService.checkProductRatingUser(
+      id,
+      user.id,
+    );
+    if (!checkUser) {
+      throw new ForbiddenException(['forbidden']);
+    }
     return await this.productRatingsService.updateProductRating(
       productId,
       id,
@@ -48,10 +65,19 @@ export class ProductRatingsController {
   }
 
   @Delete(':productId/:id')
+  @Roles(Role.Admin, Role.Manager, Role.Sales, Role.Customer)
   async deleteProductRating(
+    @ReqUser() user: User,
     @Param('productId', ParseIntPipe) productId: number,
     @Param('id', ParseIntPipe) id: number,
   ): Promise<void> {
+    const checkUser = await this.productRatingsService.checkProductRatingUser(
+      id,
+      user.id,
+    );
+    if (!checkUser) {
+      throw new ForbiddenException(['forbidden']);
+    }
     await this.productRatingsService.deleteProductRating(productId, id);
   }
 }
